client/carousel: add carousel and component for creatCharacterPage
